+++ c++/pod5_format_pybind/api.h
@@ -338,7 +338,8 @@ inline pod5::ReadData make_read_data(
     py::array_t<float, py::array::c_style | py::array::forcecast> const & predicted_scaling_shift,
     py::array_t<std::uint32_t, py::array::c_style | py::array::forcecast> const &
         num_reads_since_mux_change,
-    py::array_t<float, py::array::c_style | py::array::forcecast> const & time_since_mux_change)
+    py::array_t<float, py::array::c_style | py::array::forcecast> const & time_since_mux_change,
+    py::array_t<float, py::array::c_style | py::array::forcecast> const & open_pore_level)
 {
     auto read_ids = reinterpret_cast<pod5::Uuid const *>(read_id_data.data(0));
     return pod5::ReadData{
@@ -360,7 +361,8 @@ inline pod5::ReadData make_read_data(
         *predicted_scaling_scale.data(row_id),
         *predicted_scaling_shift.data(row_id),
         *num_reads_since_mux_change.data(row_id),
-        *time_since_mux_change.data(row_id)};
+        *time_since_mux_change.data(row_id),
+        *open_pore_level.data(row_id)};
 }
 
 inline void FileWriter_add_reads(
@@ -387,6 +389,7 @@ inline void FileWriter_add_reads(
     py::array_t<std::uint32_t, py::array::c_style | py::array::forcecast> const &
         num_reads_since_mux_changes,
     py::array_t<float, py::array::c_style | py::array::forcecast> const & time_since_mux_changes,
+    py::array_t<float, py::array::c_style | py::array::forcecast> const & open_pore_levels,
     py::list signal_ptrs)
 {
     if (read_id_data.shape(1) != 16) {
@@ -422,7 +425,8 @@ inline void FileWriter_add_reads(
             predicted_scaling_scales,
             predicted_scaling_shifts,
             num_reads_since_mux_changes,
-            time_since_mux_changes);
+            time_since_mux_changes,
+            open_pore_levels);
 
         throw_on_error(w.add_complete_read(read_data, signal_span));
     }
@@ -452,6 +456,7 @@ inline void FileWriter_add_reads_pre_compressed(
     py::array_t<std::uint32_t, py::array::c_style | py::array::forcecast> const &
         num_reads_since_mux_changes,
     py::array_t<float, py::array::c_style | py::array::forcecast> const & time_since_mux_changes,
+    py::array_t<float, py::array::c_style | py::array::forcecast> const & open_pore_levels,
     py::list compressed_signal_ptrs,
     py::array_t<std::uint32_t, py::array::c_style | py::array::forcecast> const & sample_counts,
     py::array_t<std::uint32_t, py::array::c_style | py::array::forcecast> const &
@@ -511,7 +516,8 @@ inline void FileWriter_add_reads_pre_compressed(
             predicted_scaling_scales,
             predicted_scaling_shifts,
             num_reads_since_mux_changes,
-            time_since_mux_changes);
+            time_since_mux_changes,
+            open_pore_levels);
 
         throw_on_error(w.add_complete_read(read_data, signal_rows, signal_duration_count));
     }